% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateObject.R
\name{annotateObject}
\alias{annotateObject}
\alias{objectAnnotation}
\alias{setAnnotation}
\title{Add calcite annotation}
\usage{
annotateObject(x, title, description, maintainers, species, genome, origin)

objectAnnotation(x)

setAnnotation(x, annotation)
}
\arguments{
\item{x}{An \linkS4class{Annotated} object.}

\item{title}{String containing the object's title.}

\item{description}{String containing the description of the object.}

\item{maintainers}{List containing the identity of the maintainers.
Each entry may be a \code{\link{person}}-formatted string, or a list containing the \code{name} and \code{email} strings.

Alternatively, a character vector of \code{\link{person}}-formatted strings.}

\item{species}{An integer vector of NCBI taxonomy IDs for the species relevant to the data in \code{x}.}

\item{genome}{List describing the genomes involved in constructing \code{x}.
Each entry should be a list containing:
\itemize{
\item \code{id}, a string containing the genome build ID (e.g., hg38, NCBIm37)
\item \code{type}, a string specifying the type of ID (e.g., UCSC, Ensembl)
}
Check out \url{https://artifactdb.github.io/calcite-schemas/array/v1.html#allOf_i0_genome} for details.}

\item{origin}{List describing the origin of \code{x}.
Each entry should be a list containing:
\itemize{
\item \code{source}, a string describing the source repository, e.g., PubMed, GEO, ArrayExpress.
\item \code{id}, a string containing an identifier within \code{source}.
}
Check out \url{https://artifactdb.github.io/calcite-schemas/array/v1.html#allOf_i0_origin} for details.}

\item{annotation}{List of calcite-relevant metadata.}
}
\value{
For \code{annotateObject}, \code{x} is returned with extra fields in its \code{\link{metadata}}.

For \code{objectAnnotation}, a list of calcite-relevant metadata is returned.

For \code{setannotation}, \code{x} is returned after replacing the calcite-relevant metadata with \code{annotation}.
}
\description{
Add calcite's required annotation to a Bioconductor \linkS4class{Annotated} object.
}
\examples{
library(S4Vectors)
df <- DataFrame(X=1:10, Y=LETTERS[1:10], Z=factor(letters[1:10]))

df <- annotateObject(df,
    title="FOO",
    description="I am a data frame",
    maintainers="Aaron Lun <infinite.monkeys.with.keyboards@gmail.com>",
    species=9606,
    genome=list(list(id="hg38", source="UCSC")),
    origin=list(list(source="PubMed", id="123456789"))
)

anno <- objectAnnotation(df)
str(anno)

anno$maintainers <- c(anno$maintainers, 
    list(list(name="Darth Vader", email="vader@empire.gov")))
df <- setAnnotation(df, anno)

}
\author{
Aaron Lun
}
