% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalciteArray.R
\docType{class}
\name{CalciteArray}
\alias{CalciteArray}
\alias{CalciteArray-class}
\alias{CalciteMatrix-class}
\alias{matrixClass,CalciteArray-method}
\alias{DelayedArray,CalciteArraySeed-method}
\alias{CalciteArraySeed-class}
\alias{loadArray}
\alias{CalciteArraySeed}
\title{The CalciteArray class}
\usage{
CalciteArraySeed(id)

CalciteArray(id, ...)

\S4method{DelayedArray}{CalciteArraySeed}(seed)
}
\arguments{
\item{id}{String containing the calcite identifier, or a list containing the unpacked components of an identifier (from \code{\link{unpackID}}).
For \code{CalciteArray}, this can also be a CalciteArraySeed object.}

\item{...}{Further arguments to pass to the CalciteArraySeed constructor.}

\item{seed}{A CalciteArraySeed object.}
}
\value{
The \code{CalciteArraySeed} constructor returns a CalciteArraySeed object.

The \code{CalciteArray} constructor and the \code{\link{DelayedArray}} method return a CalciteArray object (or a CalciteMatrix, for two dimensions).
}
\description{
The CalciteArray is a \pkg{DelayedArray} wrapper around a on-disk array downloaded from the calcite backend.
Specifically, it remembers the zircon ID that was used to download the array, which can be used to avoid a redundant write and upload in \code{\link{saveObject}} and \code{\link{uploadDirectory}}, respectively.
The actual heavy-lifting is done by forwarding all operations to the internal \pkg{DelayedArray}-compatible seed object,
which may be any of the usual classes, e.g., a HDF5ArraySeed, H5SparseMatrixSeed or any of its delayed operations.
}
\details{
If \code{id} uses the \code{latest} version alias, this is automatically resolved to a full version by the \code{CalciteArraySeed} constructor.
This ensures that the version is explicitly pinned in downstream applications and when saving to file.
}
\examples{
\dontshow{fun <- calcite:::.configure_cache()}
id <- "test:my_first_sce/assay-1/matrix.h5@v1"
mat <- CalciteArray(id)
mat
\dontshow{fun()}

}
