% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleObject.R
\name{exampleObject}
\alias{exampleObject}
\alias{exampleID}
\title{Example object and IDs for testing}
\usage{
exampleObject()

exampleID()
}
\value{
For \code{exampleObject}, a \linkS4class{DFrame} is returned with annotation set by \code{\link{annotateObject}}.

For \code{exampleID}, a string is returned containing an example ID.
}
\description{
Make some example objects for testing.
}
\examples{
exampleObject()
exampleID()

}
\author{
Aaron Lun
}
