% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restURL.R
\name{restURL}
\alias{restURL}
\title{Get/set the calcite REST URL}
\usage{
restURL(url)
}
\arguments{
\item{url}{String containing the REST URL to use.}
}
\value{
If \code{url} is not supplied, the currently set URL is returned.

If \code{url} is supplied, it is used to set the URL, and the \emph{previously} set value of the URL is returned.
}
\description{
Pretty much as the title says.
}
\examples{
restURL()

old <- restURL("https://new.url.com")
restURL()
restURL(old)

}
\author{
Aaron Lun
}
