% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveObject.R
\name{saveObject}
\alias{saveObject}
\title{Save an object to file}
\usage{
saveObject(x, dir, path)
}
\arguments{
\item{x}{A supported Bioconductor object, with annotation added by \code{\link{annotateObject}}.}

\item{dir}{String containing a path to a staging directory.}

\item{path}{String containing the relative path inside the staging directory in which to save the object's contents.
This should not be nested inside any subdirectories created by previous \code{saveObject} calls.}
}
\value{
\code{x} is saved to the specified location, and \code{NULL} is invisibly returned.
}
\description{
Save an object to a staging directory in preparation for upload.
Multiple objects may be saved inside the same staging directory,
though users should avoid saving an object inside another object's subdirectory (created by a previous \code{saveObject} call).
}
\examples{
# Making an exciting DataFrame.
library(S4Vectors)
df <- DataFrame(X=1:10, Y=LETTERS[1:10], Z=factor(letters[1:10]))
df$AA <- DataFrame(foo=runif(10), bar=rnorm(10))

df <- annotateObject(df,
    title="FOO",
    description="I am a data frame",
    maintainers="LTLA <infinite.monkeys.with.keyboards@gmail.com>",
    species=9606,
    genome=list(list(id="hg38", source="UCSC")),
    origin=list(list(source="PubMed", id="123456789"))
)

# Staging the DataFrame.
tmp <- tempfile()
dir.create(tmp)
saveObject(df, tmp, "df1")

list.files(tmp, recursive=TRUE)

}
\seealso{
\code{\link{annotateObject}}, to add the mandatory annotation to all objects.

\code{\link{uploadDirectory}}, to upload all objects to the calcite store.
}
\author{
Aaron Lun
}
