% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveObject.R
\name{saveObject}
\alias{saveObject}
\title{Save an object to file}
\usage{
saveObject(x, dir, path)
}
\arguments{
\item{x}{A supported Bioconductor object, with annotation added by \code{\link{annotateObject}}.}

\item{dir}{String containing a path to a staging directory.}

\item{path}{String containing the relative path inside the staging directory in which to save the object's contents.
This should not be nested inside any subdirectories created by previous \code{saveObject} calls.}
}
\value{
\code{x} is saved to the specified location, and \code{NULL} is invisibly returned.
}
\description{
Save an object to a staging directory in preparation for upload.
Multiple objects may be saved inside the same staging directory,
though users should avoid saving an object inside another object's subdirectory (created by a previous \code{saveObject} call).
}
\examples{
# Making an example DataFrame:
df <- exampleObject()
df
str(objectAnnotation(df))

# Saving it to a directory:
tmp <- tempfile()
dir.create(tmp)
saveObject(df, tmp, "my_first_df")

list.files(tmp, recursive=TRUE)

}
\seealso{
\code{\link{annotateObject}}, to add the mandatory annotation to all objects.

\code{\link{uploadDirectory}}, to upload all objects to the calcite store.
}
\author{
Aaron Lun
}
